Extract shared ScrollableCarrousel component
